This is the first commit after creating the repo.
I created the maven project using the archetype in get started with javafx in Maven.

I modified the pom.xml to use 23.0.1 to be the compiler version.
I created a property for saving the javafx version that will be unified across
all javafx dependencies that we use, starting with javafx.controls.

I used the plugin version 0.0.8, since this worked, 0.0.13 refused to work
Maven tells me that It could not be found.