Big Advancemnts in the server side.
- Changed the pom.xml in the parent project so that the mysql scope is runtime only.

- Modifed the server-side/.gitignore so that it ignored the properties file that contains the
URL, username, and password that are used to connect to the database.

- Modified the App.java to test for the connection with the database at the server start, so that
I will abort if it could not establish a connection.

- Implemented the initial version of the AdminServerAPI and ClientServerAPI classes.

- Wrote down the classes package which contains classes that are not related to any specific package.
It contains the AdminSessionData.java and ClientSessionData.java which are send to the client when
it requests a login.

It Contains the AdminToken and ClientToken which are used by the server to authenticate and authorize
the access of the user at each request.

It Contains the RequestResult class which is used to wrap the response of the request that the client requests.

If the object is null, then the database is down or an unsolved problem happened at the server side.

- Added topcontrollers package which contains the DaoAdminController and DaoClientController which
is used to separate the API classes from the data access layer, they also validate the client at each
request using the Token Validator dao class.

- I added the dao package which represents the data access layer.
it contains the users dao, the status dao, the service dao, and the token validator dao.

- I added the dbconnections package which currently contain ConnectionManager class, you
use this class to get a connection with the database.

- I modifed the module-info.java to be able to use java.sql and java.naming.

- I implemented some of the models/entities that we have in the system.